requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
prisma==0.12.0
httpx==0.26.0
PyJWT==2.8.0
google-generativeai==0.3.2
Pillow==10.2.0
supabase==2.3.4
python-multipart==0.0.6
